test: refactor core-graphql
