+++ packages/core-graphql/__tests__/api/transaction.test.js
@@ -1,12 +1,13 @@
 const app = require('../__support__/setup')
 const utils = require('../__support__/utils')
+require('@arkecosystem/core-test-utils/lib/matchers')
 
 let genesisBlock
 
 beforeAll(async () => {
   await app.setUp()
 
-  genesisBlock = require('../__fixtures__/genesisBlock.json')
+  genesisBlock = require('@arkecosystem/core-test-utils/config/testnet/genesisBlock.json')
 })
 
 afterAll(() => {
@@ -19,10 +20,10 @@ describe('GraphQL API { transaction }', () => {
       const query = `{ transaction(id:"${genesisBlock.transactions[0].id}") { id } }`
       const response = await utils.request(query)
 
-      utils.expectSuccessful(response)
-      utils.expectResource(response)
+      expect(response).toBeSuccessfulResponse()
 
       const data = response.data.data
+      expect(data).toBeObject()
       expect(data.transaction).toBeObject()
       expect(data.transaction.id).toBe(genesisBlock.transactions[0].id)
     })
