+++ packages/core-graphql/__tests__/api/transactions.test.js
@@ -1,5 +1,6 @@
 const app = require('../__support__/setup')
 const utils = require('../__support__/utils')
+require('@arkecosystem/core-test-utils/lib/matchers')
 
 beforeAll(async () => {
   await app.setUp()
@@ -15,10 +16,10 @@ describe('GraphQL API { transactions }', () => {
       const query = '{ transactions(filter: { fee: 0 }) { id } }'
       const response = await utils.request(query)
 
-      utils.expectSuccessful(response)
-      utils.expectResource(response)
+      expect(response).toBeSuccessfulResponse()
 
       const data = response.data.data
+      expect(data).toBeObject()
       expect(data.transactions.length).toBe(100) // because of default limit = 100
     })
 
@@ -26,10 +27,10 @@ describe('GraphQL API { transactions }', () => {
       const query = '{ transactions(filter: { fee: 987 }) { id } }'
       const response = await utils.request(query)
 
-      utils.expectSuccessful(response)
-      utils.expectResource(response)
+      expect(response).toBeSuccessfulResponse()
 
       const data = response.data.data
+      expect(data).toBeObject()
       expect(data.transactions.length).toBe(0)
     })
   })
